Removed TF-IDF calculations due to issues, will be readded later.
Initially implemented a script to calculate Term Frequency (TF), Inverse Document Frequency (IDF), and TF-IDF weights for improved search relevance, but this file was removed due to unresolved issues in the code. The intent was to enhance the search engine's ability to evaluate and rank documents based on keyword significance.